requirements.txt
Django==2.2.1
django-admin-interface==0.19.2
django-admin-tools==0.9.2
django-colorfield==0.7.2
django-crispy-forms==1.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
djangorestframework==3.9.4
Pillow==9.2.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
